Use exact `purescript` modules
The `purescript` package follows this pattern where modules re-export a
bunch of other modules. This makes it incredibly hard to know what
you're actually using from any one module. In our case, we don't really
know how dependent on the CST we really are.

Turns out, we only really care about the types. We have a handful of
dependencies on the other parts of the CST, but nothing too major. This
should provide feedback into how to split out the CST stuff from the
`purescript` package.