Use `purescript-cst` instead of `purescript`
The `purescript` package brings with it too many dependencies. We don't
want to build the networking stuff as that conflicts with the `dhall`
package`. We don't care that it does things with `align` because we're
not even privvy to that code. We don't want to have to solve a bunch of
transitive dependencies for code paths we never execute.

Since `stack` doesn't have a way to specify that we only want the
library of the `purescript` package (you have to build the library and
executables with `stack`), and we don't want to switch to a completely
different package manager right now, we do the next best thing and
depend on a smaller collection of files spun out from the `purescript`
package.

There's a fork of the `purescript` package that splits out the CST stuff
that we use directly. This is all the dependencies we need, so we depend
on it directly. Hopefully, we can have this fork merged upstream so we
don't have to deal with that fork per each version change.

In any case, the change to us is mostly transparent. We see a noticable
decrease in clean build times though, so this is a huge win!