Remove configuration
The promise of https://gitlab.com/joneshf/purty/issues/18 never came to
fruition.

We thought that we'd want to configure all sorts of things, but the way
this project evolved, we want to minimize how much stuff actually is
configured.

We also thought that we'd be able to hype Dhall as a configuration
language, but we don't really get any of the benefits of Dhall with our
configuration as it was.

We also didn't think about the out of band information that a
configuration file has on a CLI. Since the config file is something that
provides default values to the CLI, it can lead to the behavior of the
CLI not being clear.

We also don't want to deal with compilation times taking forever, or
having to compile HTTP stuff. We're trying to trim our dependencies
nowadays, and `dhall` is another big one.

For all these reasons, we drop the configuration for now. We _might_
come back to it at some point, but it's unlikely.