Statically link binary
We've gone a long time without the ability to statically link the
binaries. This was mostly because our transitive dependencies (at least
`purescript`, maybe also `dhall`) brought in `zlib` and other packages
that are notoriously hard for Haskell to statically link.

Now that we no longer have these dependencies, we can statically link
fairly easily! It's actually really easy to do; just one flag. There are
some other flags around static linking and optimization we should
probably do. We can do that in a follow-up commit though.

The difference in file size between dynamic and static linking appears
to be 17M vs. 19M, respectively. Not enough to warrant worrying about.