Move more behavior into `Args.withInput`
We want to see about formatting directories. That seems like it's going
to be kind of hard while we have `Args.withInput` passing back what
should be one file.

Instead of forcing ourselves into a pattern that might not be beneficial
for us, we inline the code and start anew. Once we've implemented the
directory support, we can look back on this.