Walk a directory and format all `.purs` files
Now that we've got everything setup for it, we can format an entire
directory. The output argument is less useful for a directory if it's to
STDOUT as we end up with a bunch of modules all writtein to STDOUT with
no way to distinguish modules.

We should come back and think about how we might want to handle
directories a bit better. Maybe we follow `elm-format`s approach and
make the whole CLI interactive? Maybe we do something else?