Use `pathwalk` instead of `path-io`
`pathwalk` is equivalent to `path-io` for what we want, but doesn't fail
to build with bazel. We switch to it and use it instead of `path-io` for
the time being.

This isn't a fix, this is a hack we do since we're unfamiliar with bazel
and not sure how to fix the underlying problem. If we figure that out,
we might consider changing back to `pathwalk`.