Bootstrap `bazel` with `make`
We can't assume we'll always have `bazel` available, but we'll be pretty
confident that we'll have `make` and `curl` available. With these tools,
we can bootstrap the `bazel` binary, and use that to build out the rest
of the system.

We're doing this small step first to flesh out any incorrect assumptions
with what we're thinking. We might end up having to iterate on this.