Update to the latest version of GNUMake
Version 4.2.1 is no longer available through `chocolatey`. The version
that's available is now 4.3. This has actually been broken for a while,
but we didn't deal with it since the cache was not invalidated until we
changed the `appveyor.yml` file.

We might consider a more reliable way to ensure we have the correct
version of `make` installed.