Try setting the (implicit) Python toolchain
When using GHC bindist with `rules_haskell`, it assumes there's the
correct Python installed somewhere outside of its scope. It uses
whatever Python toolchain is available in the global scope. We
explicitly set the toolchain to be these values that we _think_ should
exist. Since each platform has different locations for its Python 3
value, we have to select based on the platform we're on.

It was incredibly hard to figure out how to set platform specific values
in bazel. There's a good chance that this isn't correct.