Move `ormolu` script to `Makefile`
We're doing more stuff in the Makefile, so we move the actual linting to
there as well. This terseness isn't out of laziness, rather that there
are just fewer things to do in a `Makefile` compared to a shell script.

The way we're handling the error isn't great, but it'd be better if
`ormolu` could give a useful error message when it fails. Maybe
something to request upstream?