Remove `bytestring` dependency from library
Similar to the change we did for removing `text`, we remove the
`bytestring` dependency from the library. We still have it
transitively (it's part of the GHC toolchain), but we're not depending
on it explicitly.

It doesn't look like we can drop the `bytestring` dependency in the test
suite right now. That's okay. We can come back to this at a later date
and maybe be able to drop it.