Split handling of input per command
We want to allow different help text for different commands. It's more
work than necessary to keep these things the same, so we split them up.
The `format` command can have its help text and the `validate` command
can have its help text.

Really, it's not right that `Input` even exists, as it has only just now
become useable. We're abusing `Input` for `Format` because there's a
case that doesn't make sense: `Format' InputSTDIN Write`. We're not
going to write to STDIN (though you techincally can), and indeed, we
don't. We shouldn't let that be a possible state, yet it is.

What we should do, is split out the different formatting states, so
things can be handled appropriately.