Try patching the encoding error handling
`open` defaults to raising exceptions when there is an encoding error.
Since this code exists to `sed` something (and it doesn't seem to care
what the non-regex bytes are), we ought to be able to escape the unknown
bytes to the private use area, then unescape them when writing the file
back out.

There's more information about the different ways to handle errors under
the `error` section:
https://docs.python.org/3/library/functions.html#open.