Try setting `BAZEL_USE_CPP_ONLY_TOOLCHAIN`
We've been running into a failure for a while on macOS:
```
Setup.hs: Missing dependency on a foreign library:
* Missing (or bad) header file: primitive-memops.h
This problem can usually be solved by installing the system package that
provides this library (you may need the "-dev" version). If the library
is already installed but in a non-standard location then you can use the
flags --extra-include-dirs= and --extra-lib-dirs= to specify where it
is.If the library file does exist, it may contain errors that are caught
by the C compiler at the preprocessing stage. In this case you can
re-run configure with the verbosity flag -v3 to see the error messages.
If the header file does exist, it may contain errors that are caught by
the C compiler at the preprocessing stage. In this case you can re-run
configure with the verbosity flag -v3 to see the error messages.
```

Some searching shows that this can be caused by not having the correct C
toolchain setup: https://stackoverflow.com/a/25905228/1549047. Much like
the fixes we did for the Windows build, it looks like we'll have to
setup the macOS `cc_toolchain` to point to the right spot.

The troubleshooting steps for `rules_haskell` mention the
`BAZEL_USE_CPP_ONLY_TOOLCHAIN` environment variable and state that it's
used to choose the correct `cc_toolchain`:
https://github.com/tweag/rules_haskell/tree/9b88fc52fab4e7d9e6aee25a6b5585ed9355bf3b#macos-error-developer_dir-not-set.

We set it in hopes that it fixes all of our woes.

Since the `.travis.yml` file is used by both the Linux and macOS builds,
this might break the Linux build. If it does, we can probably figure
some way to only set this environment variable on macOS builds.

Alternatively, it could make the Linux builds _more_ hermetic. Guess
we'll have to see what happens when it builds.