Pull Haskell sources up to variables
The approach we're planning to take with `hlint` is to reference each
file individually. It doesn't seem possible to reference the sources if
they're within a rule. Maybe there's some way to do it (which would make
this easier), but it's not clear how.

If we figure another way to do this, we can always move to that later.