Run `hlint` through `bazel`
We can download an `hlint` binary for each of the platforms and choose
which one to use in an `alias`.

This goes against full hermeticity, but we don't want to spend time
building `hlint`.

We also start doing things with tags. In the kubernetes, prometheus, and
terraform world, labeling/tagging is incredibly useful after the fact.
There's hints that it can be useful as well in the bazel world. So, we
start tagging more stuff. We hope to use this to make the make rules a
little easier.

We made use of tagging already by creating two different `test_suite`s:
one for hlint specifically, and one for all linting. Hopefully, the make
setup will be able to use these easily.