Run `ormolu` through `bazel` instead of `make`
Now that we've got `ormolu` setup in `bazel`, we can use it instead of
using `make`. This is the last vestiage of `stack` as well. Which means
we can dump quite a bit of stuff now that we're not reliant on `stack`
directly in our build; we still transitively use `stack` from `bazel`.

We'll think about changing how we deal with `stack` from `bazel` when
the time comes.