Run npm acceptance test in bazel
If we bring in a toolchain for node.js, we can run the acceptance tests
through bazel!

There's some fun stuff with setting up the different files this test
needs, but it seems to work out alright. The real test will be whether
things work on our different CI services.