Don't generate a file
We can use the string directly instead of generating a file.

We already have the value we want, we don't need to write it out to a
file, then read it back into a file later in the tests. We move to just
giving the string to the test and calling it a day.