Template the `package.json` with `bazel`
We can create the `package.json` from `bazel` now.

Similar to the move for `bintray.json`, we can generate the JSON file
from `bazel`. This allows us to remove our last `dhall-to-json`
dependency.

We'd like to use `rules_nodejs`'s `pkg_npm` rule. Unfortunately, we have
a `prepack` script that cannot work properly. It needs to make
directories, but it's run in a location where it doesn't have write
access.

We could try a different packaging approach: one package per platform
and optional dependencies on a main package. But, that's a lot of work
for addressing this one thing.