Remove platform argument to `make`
We can figure out the platform ourselves.

We don't need to ask the consumer of our `Makefile` to supply the
platform they're running it on. We have access to enough information to
figure that out on our own.

Because we don't have to ask for the platform, we can clean up the
dependencies a bit. We can eschew the need for a `bootstrap` rule. We
already know which binaries we want, so we can just download them
whenever we need them.