Pass `make` arguments directly
Environment variables are not the same as `make` variables.

This is a thing that catches us out from time to time. We have to
explicitly pass variables in instead of assuming they'll be available
from the environment.

FWIW, this is what we probably should have done from the start. Instead
of using out of band information like the environment variables.