Run golden tests in `bazel`
We can run everything in `bazel` using the actual `purty` binary.

We initially started running tests in Haskell because it was an easy
thing to do. We don't actually have to run them in Haskell in order to
get the confidence we want. In all actuality, we have more confidence
now because we're using the actual interface to `purty`.

We set up a couple of macros to make running these rules a little easier
and let `bazel` take care of all the details. The best part about this
is that we get to benefit from `bazel`'s caching and file tracking to
make things work a little better.