Remove Haskell test suite
Now that `bazel` is running the show, we don't need this suite.

We can drop not only the suite, but also a few dependencies. This works
out well because it means there's less stuff to build and maintain.

We should feel free to add a test suite for Haskell stuff, but it should
test Haskell-specific stuff that cannot be tested external to the
interface (the `purty` binary).