Pass the binary into the macro
We don't want to have the label down inside the macro.

Macros don't provide great error messages, so we don't want to have to
hunt down what's going wrong. Instead of trying to dig out what the
issue is, we move it further up so when the target changes, we're aware
of it where it's an issue.

It's still not perfect, but at least it's more visible.