Require a name and make a test suite for hlint
It's best practice for every macro to require a name.

This might be showing why it's a best practice. With this name, we can
create the test suite in the macro. This keeps a bunch of the logic in
the macro instead of splitting it across two locations in the build
file.

We also switch to naming the test directly in the `//:lint`
`test_suite`. This should make it a bit easier to not forget to update
things when they change. Tags are useful, but they don't require
existence of targets the same way that using a label does.