Provide a template rule that works per package
We want to move the templated targets to their own package.

We can't really move the templates without breaking something since
`ctx.file.template.path` gives the full path from the root. What we want
is `ctx.file.template.basename` (so it's local to the package). If we
were to make that change directly, it'd break any templates not directly
in a package in the opposite way (they'd all be at the root).

We choose to copy the rule and change how we make the output file. This
is a bit verbose, but not a big deal.