Move `bintray.json` to its own package
We follow the pattern to move this template to its own package.

We move most of the tarball stuff to this package as well. It's specific
to how TravisCI works, so it should be there. The one target we don't
remove is the Windows tarball. This is used by AppVeyor. We should make
something specific to AppVeyor so it can work similarly.