Try moving the flag to `bazel test`
We seem to have somehow acquired a `@local_jdk` dependency.

Something weird is going on where after moving `//:purty` to
`//lib:purty`, we're trying to do things with `@local_jdk`. It's unclear
how or why this is happening. It's also unclear how to stop it from
happening. The part that's hard is that we're acquiring this dependency
from `//:test-acceptance-binary`, but we're not supposed to be running
this command because it has a tag that should exclude it.

We move the `--test_tag_filters` flag to `bazel test` in hopes that it
actually filters things out.