Move version to an internal package
We don't really want version stuff to be specific to `purty`.

We'd like to keep all the versioning stuff as its own thing. Instead of
mixing it with the `purty` binary, we want it to be thought of as its
own stuff. We move it to an `internal` directory in an attempt to
signify that this isn't stuff we expect anyone to depend on, or even
care that it exists.