Move `SourceRange` module to its own package
We move stuff very close to `purescript-cst` to its own package.

This module only exists here because `purescript-cst` doesn't provide
these values. It probably doesn't provide them because it has no
immediate use for them.

We move this module to the `pkg` directory because we don't really
expect it to be something internal-only. This should really be in
upstream because the values only operate on stuff in `purescript-cst`.

We should see if upstream would accept these values so we don't have to
maintain them. Not a huge deal if not. But it'd be nice to not have to
maintain them.