Move formatting to its own package
We separate out the bulk of formatting to its own package

Turns out we actually weren't doing all that much in the core of the
program. The remaining modules are not actually the formatting stuff.
They're part of the actual program that we are trying to make. It's nice
that we can split out packages in a way that makes that clearer.