Move `src` package to `cmd/purs-format`
We experiment with a different structure.

Now that we're no longer constrained by thinking only about Haskell
stuff, we can restructure the files. We're taking a page from golang on
this one. The binary we're building ends up living in `cmd/purs-format`.

With this change, it becomes easier to see that we might make more than
one binary. That is the plan, eventually. So this is forward movement
in that world.

One other important part of this is that we're starting to rename to
`purs-format`. This name is chosen to fall in line more with stuff
outside of the PureScript ecosystem. The name `purty` was cute, but it
didn't inspire immediate knowledge of what the tool was for or would do.