Clean up error rendering
The error rendering is mostly noise that muddies up the signal.

Handling errors in Haskell is incredibly verbose if we're to do it in a
way that doesn't involve introducing accidental complexity of monad
transformers or some such.

To make things a bit more palatable, we pull out the rendering of errors
to their own values. We compose things back together at the top in a way
that makes it a bit easier to see what's up.