Handle trailing module comments correctly
We don't want to sqaush all trailing module comments into one.

The way we deal with trailing comments is not quite right. Part of this
is because we didn't understand how the PureScript parser works when
writing this code initially. Part of this is because trailing comments
are much more complex than leading comments.

The way (we think) the parser works is that only things that immediately
follow are parsed as trailing comments. Things that look like they
should be trailing comments aren't necessarily trailing comments. In
particular, if there's a bunch of line comments after something, the
first comment is a trailing comment while the rest are leading comments
on the next thing. But, only sometimes! If there's nothing after that
could be a leading comment, they're all trailing comments.

The long and short is that trailing comments add a whole different level
of complexity. Whether that complexity is worth it is up for debate.

In any case, we handle trailing comments on modules different from other
trailing comments. When we get to the end of a module, we process the
"comments" to see if there are any actual comments. That's the other bit
of complexity here. A `Language.PureScript.CST.Comment _` is not
necessarily a comment. It could be some number of spaces, or a line
feed, or a carriage return and line feed. We don't care about anything
but comments, so we have to process out just those pieces.

If there are any actual comments, we put a newline between each and
format them at the end of the module. It's just that simple!