Bump `rules_haskell` to 0.13
We worked around the sub-package issue, so we can use a newer version.

This version brings a welcomed fix: better caching of the stack setup.
This was mentioned in the GitHub issue:
https://github.com/tweag/rules_haskell/issues/1293. It seems like using
the `stack_snapshot_json` attribute might address all of the issues
rraised in that GitHub issue. If it only addresses the local
intermittent checks it would be a win by itself.

We'll try it out, and see where we get.