Use the correct `bazel` binary for `ibazel`
We don't want to use whatever `bazel` is on the `$PATH`.

If we don't specify `-bazel_path`, `ibazel` will use whatever `bazel` it
finds on the `$PATH`. That means that the call to `ibazel` could use a
version of `bazel` that would invalidate all of the building we do with
our project-local version of `bazel`.

To make things a bit more consistent, we specify the `-bazel_path` as
the `bazel` we use all throughout the rest of the `Makefile`.