Downgrade the image on AppVeyor
We can no longer use the latest image on AppVeyor.

On 2020-11-14, AppVeyor updated the version of .NET to .NET 5:
https://www.appveyor.com/blog/2020/11/14/dotnet-5-and-dotnet-core-sdk-version-pinning/.
That update caused the C compiler toolchain to move elsewhere that is no
longer picked up by bazel.

Per the recommendations in that blog post, we drop down to the last
image. This image should use Visual Studio 2019 version 16.7.6 without
.NET 5.

Hopefully, this works the first time out.