Change each goal/non-goal to a heading
We want to be able to link directly to a goal/non-goal.

Although these are lists, most markdown processors don't automatically
create links for each element in a list. In particular, GitLab doesn't
make a link for each of the list items.

Since we want links, we have at least two options:
1. We could manually make anchor tags with our own links.
2. We could change the markdown we use to something that generate links.

We choose to do the second option. Maybe it's more correct semantically
to use headings and we should have done this from the start?

In any case, this should give us the ability to link directly to an
individual goal/non-goal when we want to.