Bump `bazel` to 3.7.0
We update to a newer version of `bazel`.

We don't go to 4.0.0 because that just came out a couple of days ago,
and living on the bleeding edge is not a lot of fun. Given that it took
multiple months to update `rules_haskell`, it seems reasonable to wait a
while before updating across a major version change of `bazel`.